move dependencies to top level, and add design note
