+++ antarest/study/web/directory_blueprint.py
@@ -17,8 +17,8 @@ from typing import List
 from fastapi import APIRouter, Depends
 
 from antarest.core.api_types import UuidStr
-from antarest.core.dependencies import DirectoryServiceDep, auth_required
 from antarest.core.utils.web import APITag
+from antarest.dependencies import DirectoryServiceDep, auth_required
 from antarest.study.model import DirectoryCreation, DirectoryMetadata, DirectoryUpdate
 
 logger = logging.getLogger(__name__)
