+++ tests/integration/login/test_auth.py
@@ -16,7 +16,7 @@ from fastapi import FastAPI
 from starlette.testclient import TestClient
 
 from antarest.core.config import Config, SecurityConfig
-from antarest.core.dependencies import get_config
+from antarest.dependencies import get_config
 
 
 @pytest.fixture
